% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getclaims}
\alias{getclaims}
\title{Get claims of a Wikidata item}
\usage{
getclaims(item, print = TRUE)
}
\arguments{
\item{item}{The Wikidata item object - from \code{\link{wdgetitem}}}

\item{print}{Logical - if \code{TRUE} (default) the claims are printed}
}
\value{
A data frame listing all claims of the item
}
\description{
Get claims of a Wikidata item
}
\examples{
\dontrun{
zapa.item <- wdgetitem(qid="q144786")
getclaims(zapa.item)
zapa.claims <- getclaims(item=zapa.item, print=FALSE)
}
}
\seealso{
\code{\link{wdgetitem}} for general information about the item, \code{\link{wdgetclaims}} to get claims of a wditem object by API request
}

