% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getqualifier}
\alias{getqualifier}
\title{Internal method to get and prepare the qualifiers}
\usage{
getqualifier(qualifier, type, lang)
}
\arguments{
\item{qualifier}{Raw qualifier object}

\item{lang}{Language abbreviation (ISO language codes), as string}
}
\value{
Qualifier value - object depends on qualifier type
}
\description{
Internal method to get and prepare the qualifiers
}
\seealso{
\code{\link{wdclaimcontent.int}}
}
\keyword{internal}

