% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/wdclaimcontent.R
\name{print.wdclaimcontent}
\alias{print.wdclaimcontent}
\title{Print method for wdclaimcontent}
\usage{
\method{print}{wdclaimcontent}(x, open.ext = TRUE, ...)
}
\arguments{
\item{x}{wdclaimcontent object from \code{\link{wdclaimcontent}}}

\item{open.ext}{Logical - if \code{TRUE} external sources of the claim like images or URLs are opened}

\item{\dots}{Arguments to be passed to methods}
}
\description{
Print method for wdclaimcontent
}

