% Generated by roxygen2 (4.0.1): do not edit by hand
\name{print.wdclaims}
\alias{print.wdclaims}
\title{Print method for wdclaim}
\usage{
\method{print}{wdclaims}(claim)
}
\arguments{
\item{claim}{wdclaim object from \code{\link{wdgetclaims}} or \code{\link{getclaims}}}
}
\description{
Print method for wdclaim
}

