% Generated by roxygen2 (4.0.1): do not edit by hand
\name{print.wdcontent}
\alias{print.wdcontent}
\title{Print method for wdcontent}
\usage{
\method{print}{wdcontent}(content, open.ext = TRUE)
}
\arguments{
\item{content}{wdcontent object from \code{\link{wdclaimcontent}}}

\item{open.ext}{Logical - if \code{TRUE} external sources of the claim like images or URLs are opened}
}
\description{
Print method for wdcontent
}

