% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/wdgetclaims.R
\name{print.wdgetclaims}
\alias{print.wdgetclaims}
\title{Print method for wdgetclaims}
\usage{
\method{print}{wdgetclaims}(x, ...)
}
\arguments{
\item{x}{wdgetclaims object from \code{\link{wdgetclaims}} or \code{\link{getclaims}}}

\item{\dots}{Arguments to be passed to methods}
}
\description{
Print method for wdgetclaims
}

