% Generated by roxygen2 (4.0.1): do not edit by hand
\name{print.wdgetitem}
\alias{print.wdgetitem}
\title{Print method for wdgetitem}
\usage{
\method{print}{wdgetitem}(x, ...)
}
\arguments{
\item{x}{wdgetitem object}

\item{\dots}{Arguments to be passed to methods}
}
\description{
Print method for wdgetitem
}

