% Generated by roxygen2 (4.0.1): do not edit by hand
\name{print.wdgetproperty}
\alias{print.wdgetproperty}
\title{Print method for wdgetproperty}
\usage{
\method{print}{wdgetproperty}(x, ...)
}
\arguments{
\item{x}{wdgetproperty object}

\item{\dots}{Arguments to be passed to methods}
}
\description{
Print method for wdgetproperty
}

