% Generated by roxygen2 (4.0.1): do not edit by hand
\name{query}
\alias{query}
\title{Internal method to execute API request and return the content}
\usage{
query(url, as_what)
}
\arguments{
\item{url}{API request URL}

\item{as_what}{Desired type of output, one of \code{raw}, \code{text} or \code{parsed.content}}
}
\value{
Content in specified type
}
\description{
Internal method to execute API request and return the content
}
\keyword{internal}

