% Generated by roxygen2 (4.0.1): do not edit by hand
\name{wdclaimcontent.int}
\alias{wdclaimcontent.int}
\title{Internal method to execute API request and process the content}
\usage{
wdclaimcontent.int(url, lang)
}
\arguments{
\item{url}{API request URL}

\item{lang}{Language abbreviation (ISO language codes), as string}
}
\value{
A list of guid, item, property, type and content
}
\description{
Internal method to execute API request and process the content
}
\seealso{
\code{\link{wdclaimcontent}}
}
\keyword{internal}

