% Generated by roxygen2 (4.0.1): do not edit by hand
\name{wdclaimcontent}
\alias{wdclaimcontent}
\title{Get content of a specific claim}
\usage{
wdclaimcontent(guid, lang = "en", print = TRUE, ...)
}
\arguments{
\item{guid}{The Wikidata claim guid, as string}

\item{lang}{Language abbreviation (ISO language codes), as string - default is \code{"en"}}

\item{print}{Logical - if \code{TRUE} (default) the claim content is printed}

\item{...}{Arguments passed to methods, e.g. \code{open.ext} - if \code{TRUE} (default) external sources of the claim like images or URLs are opened}
}
\value{
A list containing meta information of the claim and its content
}
\description{
Get content of a specific claim
}
\examples{
\dontrun{
wdclaimcontent("q206904$26C11CAB-2FEE-4A03-9774-9B70777616B3")  # wikidata item
wdclaimcontent("q52$61ba10af-455f-021c-19d5-d1a0df4d65f5")  # date/time
wdclaimcontent("q7186$5675dddd-438d-bd8d-3355-8117747e182d")  # string
wdclaimcontent("Q206904$920876CE-1C0E-40ED-BB3C-ACB2AC15870C")  # url
wdclaimcontent("q2280$91220430-4b56-8c53-d75d-ba85ce8a2629")  # geocoordinates
wdclaimcontent("q144786$25DC2C5E-D59F-4C9B-A307-1DBDF3215576")  # commons image
zapa.coa <- wdclaimcontent(guid="q144786$25DC2C5E-D59F-4C9B-A307-1DBDF3215576",
  lang="pl", print=FALSE, open.ext=FALSE)
}
}
\seealso{
\code{\link{wdgetclaims}} to get a list of all claims
}

