% Generated by roxygen2 (4.0.1): do not edit by hand
\name{wdgetclaims}
\alias{wdgetclaims}
\title{Get claims of a Wikidata item by API request}
\usage{
wdgetclaims(qid, print = TRUE)
}
\arguments{
\item{qid}{The Wikidata item id, as string (including the 'Q') or integer value (without the 'Q')}

\item{print}{Logical - if \code{TRUE} (default) the claims are printed}
}
\value{
A list with basic information about the item
}
\description{
Get claims of a Wikidata item by API request
}
\examples{
\dontrun{
wdgetclaims(144786)
zapa.claims <- wdgetclaims(qid="q144786", print=FALSE)
}
}
\seealso{
\code{\link{wdgetitem}} for general information about the item, \code{\link{getclaims}} to get claims of a wditem object from \code{\link{wdgetitem}}
}

