% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/wdgetproperty.R
\name{wdgetproperty}
\alias{wdgetproperty}
\title{Get a Wikidata property}
\usage{
wdgetproperty(pid, lang = "en", print = TRUE)
}
\arguments{
\item{pid}{The Wikidata property id, as string (including the 'P') or integer value (without the 'P')}

\item{lang}{Language abbreviation (ISO language codes), as string - default is \code{"en"}}

\item{print}{Logical - if \code{TRUE} (default) the property information are printed}
}
\value{
A vector of two: property value and property description
}
\description{
Get a Wikidata property
}
\examples{
\dontrun{
wdgetproperty(31)
prop <- wdgetproperty(pid="p31", lang="pl", print=FALSE)
}
}

