% Generated by roxygen2 (4.0.1): do not edit by hand
\name{print.wdgetclaims}
\alias{print.wdgetclaims}
\title{Print method for wdgetclaims}
\usage{
\method{print}{wdgetclaims}(x, ...)
}
\arguments{
\item{x}{wdgetclaims object from \code{\link{wdgetclaims}} or \code{\link{getclaims}}}

\item{\dots}{Arguments to be passed to methods}
}
\description{
Print method for wdgetclaims
}

