% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/wdgetitem.R
\name{print.wdgetitem}
\alias{print.wdgetitem}
\title{Print method for wdgetitem}
\usage{
\method{print}{wdgetitem}(x, ...)
}
\arguments{
\item{x}{wdgetitem object}

\item{\dots}{Arguments to be passed to methods}
}
\description{
Print method for wdgetitem
}

