% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/wdgetproperty.R
\name{print.wdgetproperty}
\alias{print.wdgetproperty}
\title{Print method for wdgetproperty}
\usage{
\method{print}{wdgetproperty}(x, ...)
}
\arguments{
\item{x}{wdgetproperty object}

\item{\dots}{Arguments to be passed to methods}
}
\description{
Print method for wdgetproperty
}

