% Generated by roxygen2 (4.0.1): do not edit by hand
\name{print.wdsearch}
\alias{print.wdsearch}
\title{Print method for wdsearch}
\usage{
\method{print}{wdsearch}(x, ...)
}
\arguments{
\item{x}{wdsearch object with search results}

\item{\dots}{Arguments to be passed to methods}
}
\description{
Print method for wdsearch
}

