% Generated by roxygen2 (4.0.1): do not edit by hand
\name{wdclaimcontent}
\alias{wdclaimcontent}
\title{Get content of a specific claim}
\usage{
wdclaimcontent(guid, lang = "en", print = TRUE, ...)
}
\arguments{
\item{guid}{The Wikidata claim guid, as string}

\item{lang}{Language abbreviation (ISO language codes), as string - default is \code{"en"}}

\item{print}{Logical - if \code{TRUE} (default) the claim content is printed}

\item{...}{Arguments passed to methods, e.g. \code{open.ext} - if \code{TRUE} (default) external sources of the claim like images or URLs are opened}
}
\value{
A list containing meta information of the claim and its content
}
\description{
Get content of a specific claim
}
\examples{
\dontrun{
wdclaimcontent("q144786$25DC2C5E-D59F-4C9B-A307-1DBDF3215576")
zapa.coa <- wdclaimcontent(guid="q144786$25DC2C5E-D59F-4C9B-A307-1DBDF3215576",
  lang="pl", print=FALSE, open.ext=FALSE)
}
}
\seealso{
\code{\link{wdgetclaims}} to get a list of all claims
}

